services/content/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2.5
pydantic-settings>=2.1
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
redis>=5.0
PyJWT>=2.8
-e ../../shared
